remove getMember/setMember

+++ DEVELOPING_OPAL/validate/src/test/java/org/opalj/fpcf/fixtures/xl/js/controlflow/intraprocedural/unidirectional/arithmetic/Mult.java
@@ -24,7 +24,7 @@ public class Mult {
         se.put("a", a);
         se.put("b", b);
         se.eval("function mult(a,b){return a * b;} var z = mult(a,b);");
-        Integer result = (Integer) se.get("z");
+        Double result = (Double) se.get("z");
         System.out.println("result: " + result);
     }
 }
